wrap apollo client setup in separate component
