+++ mobile/src/components/apollo-loader/types.ts
@@ -1,0 +1,5 @@
+import type { ReactNode } from 'react'
+
+export interface ApolloLoaderProps {
+  children: ReactNode
+}
